Added an array that hold the entire information about the game: id - from 0 to 41 arr_id is the id of ImageButton view imageButton is the view itself.
Added critical area - when the disc is falling, to prevent another play.

If column is full, it will not count as a successful play.

Added 3 Threads.
First paints the placeholder.
Second returns the white color.
Third colors the target placeholder, and saves the new color in the array.

Added lots of comments